// packages/sdk/src/generated/agreement.generated.ts
export type MeemPermissionStruct = {
  permission: PromiseOrValue<BigNumberish>;
  addresses: PromiseOrValue<string>[];
  numTokens: PromiseOrValue<BigNumberish>;
  costWei: PromiseOrValue<BigNumberish>;
  mintStartTimestamp: PromiseOrValue<BigNumberish>;
  mintEndTimestamp: PromiseOrValue<BigNumberish>;
  merkleRoot: PromiseOrValue<BytesLike>;
};

export type MeemPermissionStructOutput = [
  number,
  string[],
  BigNumber,
  BigNumber,
  BigNumber,
  BigNumber,
  string
] & {
  permission: number;
  addresses: string[];
  numTokens: BigNumber;
  costWei: BigNumber;
  mintStartTimestamp: BigNumber;
  mintEndTimestamp: BigNumber;
  merkleRoot: string;
};

export type SplitStruct = {
  toAddress: PromiseOrValue<string>;
  amount: PromiseOrValue<BigNumberish>;
  lockedBy: PromiseOrValue<string>;
};
// ...
export type SetRoleItemStruct = {
  user: PromiseOrValue<string>;
  role: PromiseOrValue<BytesLike>;
  hasRole: PromiseOrValue<boolean>;
};

export type SetRoleItemStructOutput = [string, string, boolean] & {
  user: string;
  role: string;
  hasRole: boolean;
};

export type InitParamsStruct = {
  symbol: PromiseOrValue<string>;
  name: PromiseOrValue<string>;
  contractURI: PromiseOrValue<string>;
  roles: SetRoleItemStruct[];
  maxSupply: PromiseOrValue<BigNumberish>;
  mintPermissions: MeemPermissionStruct[];
  splits: SplitStruct[];
  isTransferLocked: PromiseOrValue<boolean>;
};
// ...
export type PromiseOrValue<T> = T | Promise<T>;